Refactor pipeline installation and execution scripts
- Removed the `install_pipeline.sh` script as it was deemed unnecessary.
- Introduced `run.sh` for executing the pipeline with enhanced logging and error handling.
- Added `sbatch.template` for job submission configuration.
- Created `tsm-run` script to manage the TSM pipeline execution, including input handling and output management.
- Implemented `executor.py` to manage Dask execution configurations.
- Added `product_handle.py` for handling various file formats and data processing.
- Introduced utility functions in `utils.py` for logging and YAML file management.
- Updated `setup.sh` to streamline directory setup and configuration file generation.
- Removed the binary `point_map.fits` file from the repository.
- Added new classes and methods to improve data handling and processing within the pipeline.
