+++ config.template.yaml
@@ -1,0 +1,47 @@
+##################################################################
+# Parameters file. Should follow the syntax defined in:
+# http://docs.ansible.com/ansible/latest/YAMLSyntax.html
+# More general reference: http://www.yaml.org/spec/1.2/spec.html
+
+executor: 
+  name: "local"   # or "slurm"
+  args:
+    n_workers: 2
+    threads_per_worker: 2
+    memory_limit: "1GiB"
+  # args:
+  #   instance:
+  #     cores: 54
+  #     processes: 1
+  #     memory: "123GiB"
+  #     queue: "cpu"
+  #     job_extra_directives:
+  #       - "--propagate"
+  #       - "--time=2:00:00"
+  #   adapt:
+  #     maximum_jobs: 10
+
+inputs:
+  dataset:
+    path: "<DATASETS_DIR>/dr2"
+    columns:
+      id: "id"
+  specz:
+    - path: "<DATASETS_DIR>/specz.parquet"
+      format: "parquet"
+      columns:
+        ra: "ra"
+        dec: "dec"
+
+output_dir: "outputs"
+output_format: "csv"
+output_name: "tsm"
+
+param:
+  duplicate_criteria: "closest"
+  flux_type: "auto"
+  convert_flux_to_mag: true
+  dereddening: "sfd"
+  crossmatch:
+    radius_arcsec: 1.0
+    n_neighbors: 1
