+++ env.template.sh
@@ -1,7 +1,6 @@
 #!/bin/bash
 
 export APP_DIR="<APP_DIR>"
-export PIPELINES_DIR="<PIPELINES_DIR>"
 export DATASETS_DIR="<DATASETS_DIR>"
 
 export DASK_EXECUTOR_KEY=local
