--- training_set_maker/environment.yaml
@@ -1,16 +1,0 @@
-name: pipe_tsm
-channels:
-  - defaults
-dependencies:
-  - python=3.12
-  - pip
-  - pytables
-  - pip:
-      - PyYaml==6.0.2
-      - dask==2025.7.0
-      - distributed==2025.7.0
-      - dask-jobqueue==0.9.0
-      - lsdb==0.6.4
-      - hats==0.6.4
-      - tables_io==1.0.0
-      - h5py==3.14.0
